raspberry/requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Deprecated==1.2.10
idna==2.9
pycparser==2.20
PyGithub==1.51
PyJWT==1.7.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9
wrapt==1.12.1
